Renamed package primitives -> primitive3d
