+++ pkg/scad/primitive3d/cylinder.go
@@ -12,17 +12,20 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package primitives
+package primitive3d
 
 import (
 	"github.com/micahkemp/scad/pkg/scad/value"
 )
 
-// Cube is a cube.
-type Cube struct {
-	// Only one of these size values may be set.
-	Size    value.Float    `scad:"size"`
-	SizeXYZ value.FloatXYZ `scad:"size"`
-
-	Center value.Bool
+// Cylinder is a cylinder.
+type Cylinder struct {
+	H      value.Float `scad:"h"`
+	R      value.Float `scad:"r"`
+	R1     value.Float `scad:"r1"`
+	R2     value.Float `scad:"r2"`
+	D      value.Float `scad:"d"`
+	D1     value.Float `scad:"d1"`
+	D2     value.Float `scad:"d2"`
+	Center value.Bool  `scad:"center"`
 }
